// src/app/pages/dashboard/state/borrowing-credit-rating/models/bondIssureItemResponse.ts
export interface IBondIssureItemResponse {
  timestamp: number;
  success: boolean;
  message: string;
  data: IBondIssuerItem[];
}

export interface IBondIssuerItem {
  yearOfBondIssued: string;
  typeOfInstruments: string;
  term: string;
  couponRate: string;
  interestPayment: string;
  taxTreatment: string;
  repayment: string;
  dateOfIssue: string;
  maturityDate: string;
  platform: string;
  type: string;
  issueSize: string;
  bidsReceived: string;
  amountAccepted: string;
  greenShoeOption: string;
  greenShowOptionAmount: string;
  guaranteedByStateGovernment: string;
  guaranteeMechanism: string;
  crisil: string;
  care: string;
  icra: string;
  brickwork: string;
  "auicte/Smera": string;
  "indiaRatings&Research": string;
  linksToReports: string;
  objectOfIssue: string;
  whoCanInvest: string;
  detailsOfSubscribers: string;
  transactionAdvisors: string;
  trusteeForTheBond: string;
  registrarOfTheIssue: string;
  auditorOfIssue: string;
  legalCounsel: string;
  escrowBanker: string;
  arranger: string;
  draftInformationMemorandum: string;
  noticesFromPlatforms: string;
  others: string;
  modifiedAt: string;
  createdAt: string;
  isActive: boolean;
  _id: string;
  ulb: string;
  ulbId: string;
  __v: number;
}